% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartofriches.R
\name{cartofriches.friche}
\alias{cartofriches.friche}
\title{Renvoi la friche correpondant au site_id}
\usage{
cartofriches.friche(id_site = NULL, ...)
}
\arguments{
\item{id_site}{identifiant du site}

\item{...}{pour futurs developpements}
}
\value{
Renvoi la friche correpondant au site_id
}
\description{
Renvoi la friche correpondant au site_id
}
