% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartofriches.R
\name{cartofriches.friches}
\alias{cartofriches.friches}
\title{Retourne les friches issues de Cartofriches pour le périmètre demandé sous forme d’un dataframe}
\usage{
cartofriches.friches(
  coddep = NULL,
  code_insee = NULL,
  lon_lat = NULL,
  in_bbox = NULL,
  ...
)
}
\arguments{
\item{coddep}{Codes INSEE des départements.}

\item{code_insee}{Codes INSEE communaux ou des arrondissements municipaux.}

\item{lon_lat}{Coordonnée du point au sein de la ou des friches renvoyées c(longitude, latitude)}

\item{in_bbox}{Emprise rectangulaire sous la forme d’un vecteur c(longitude_min, latitude_min, longitude_min, latitude_max)}

\item{ordering}{Champs à utiliser pour ordonner le résultat. Default NULL}

\item{fields}{Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.}

\item{surface_max}{Surface maximale de l'unité foncière}

\item{surface_min}{Surface minimale de l'unité foncière}

\item{urba_zone_type}{Type de zone d'urbanisme}
}
\value{
Retourne les friches issues de Cartofriches pour le périmètre demandé sous forme d’un dataframe
}
\description{
Retourne les friches issues de Cartofriches pour le périmètre demandé sous forme d’un dataframe
}
\examples{
cartofriches.friches(code_insee="59350")
cartofriches.friches(coddep="59")
cartofriches.friches(in_bbox=c(3, 50, 4, 51))
}
