% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartofriches.R
\name{cartofriches.geofriches}
\alias{cartofriches.geofriches}
\title{Retourne les friches issues de Cartofriches pour le périmètre demandé sous forme d’un geodataframe}
\usage{
cartofriches.geofriches(
  coddep = NULL,
  code_insee = NULL,
  lon_lat = NULL,
  in_bbox = NULL,
  ...
)
}
\arguments{
\item{coddep}{Codes INSEE des départements.}

\item{code_insee}{Codes INSEE communaux ou des arrondissements municipaux.}

\item{lon_lat}{Coordonnée du point au sein de la ou des friches renvoyées c(longitude, latitude)}

\item{in_bbox}{Emprise rectangulaire sous la forme d’un vecteur c(longitude_min, latitude_min, longitude_min, latitude_max)}
}
\value{
Retourne les friches issues de Cartofriches pour le périmètre demandé sous forme d’un geodataframe
}
\description{
Retourne les friches issues de Cartofriches pour le périmètre demandé sous forme d’un geodataframe
}
\examples{
cartofriches.geofriches(code_insee="59350")
cartofriches.geofriches(coddep="59")
cartofriches.geofriches(in_bbox=c(3, 50, 4, 51))
}
