% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartofriches_friches_site.R
\name{cartofriches_friches_site}
\alias{cartofriches_friches_site}
\title{Retourne la friche pour l'identifiant de site demandé}
\usage{
cartofriches_friches_site(site_id = NULL)
}
\arguments{
\item{site_id}{A unique value identifying this cartofriches.}
}
\value{
Retourne la friche pour l'identifiant de site demandé
}
\description{
Retourne la friche pour l'identifiant de site demandé
}
\examples{
cartofriches_friches_site(site_id = '59002_10038')
}
