% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carofriches_geofriches.R
\name{cartofriches_geofriches}
\alias{cartofriches_geofriches}
\title{Retourne, en GeoJSON, les friches issues de Cartofriches pour la commune, le département ou l'emprise rectangulaire demandée (paramètre code_insee, coddep ou in_bbox obligatoire)}
\usage{
cartofriches_geofriches(
  coddep = NULL,
  code_insee = NULL,
  contains_geom = NULL,
  fields = NULL,
  in_bbox = NULL,
  page = NULL,
  page_size = NULL,
  surface_max = NULL,
  surface_min = NULL,
  urba_zone_type = NULL
)
}
\arguments{
\item{coddep}{Code INSEE départemental}

\item{code_insee}{Code INSEE communal ou d'arrondissement municipal (il est possible d'en demander plusieurs (10 maximum), séparés par des virgules, dans le même département)}

\item{contains_geom}{Renvoie les entités dont la géometrie contient celle précisée dans le filtre (en format GeoJSON, WKT, HEXEWKB, WKB). Exemple : /?contains_geom={'type':'Point', 'coordinates':[2.17,46.75]}}

\item{fields}{Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.}

\item{in_bbox}{Emprise rectangulaire (via Longitude min,Latitude min,Longitude max,Latitude max). L'emprise demandée ne doit pas excéder un carré de 1.0 deg. x 1.0 deg.}

\item{page}{A page number within the paginated result set.}

\item{page_size}{Number of results to return per page.}

\item{surface_max}{Surface maximale de l'unité foncière}

\item{surface_min}{Surface minimale de l'unité foncière}

\item{urba_zone_type}{Type de zone d'urbanisme}
}
\value{
Retourne les friches issues de Cartofriches pour la commune, le département ou l'emprise rectangulaire demandée (paramètre code_insee, coddep ou in_bbox obligatoire)
}
\description{
Retourne, en GeoJSON, les friches issues de Cartofriches pour la commune, le département ou l'emprise rectangulaire demandée (paramètre code_insee, coddep ou in_bbox obligatoire)
}
\examples{
cartofriches_geofriches(coddep = 59)
cartofriches_geofriches(code_insee = 59350)
cartofriches_geofriches(in_bbox = '2.827642,50.495958,3.32616,50.767734',page_size = 2)
}
