% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartofriches.R
\name{cartofriches_geofriches_dep}
\alias{cartofriches_geofriches_dep}
\title{Retourne, en GeoJSON, les friches issues de Cartofriches pour le département}
\usage{
cartofriches_geofriches_dep(
  coddep = NULL,
  fields = NULL,
  surface_max = NULL,
  surface_min = NULL,
  urba_zone_type = NULL
)
}
\arguments{
\item{coddep}{Code INSEE départemental}

\item{fields}{Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.}

\item{surface_max}{Surface maximale de l'unité foncière}

\item{surface_min}{Surface minimale de l'unité foncière}

\item{urba_zone_type}{Type de zone d'urbanisme}
}
\value{
Retourne les friches issues de Cartofriches pour le département
}
\description{
Retourne, en GeoJSON, les friches issues de Cartofriches pour le département
}
\examples{
cartofriches_geofriches_dep(coddep=59)
}
