% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{configure}
\alias{configure}
\title{configure}
\usage{
configure(...)
}
\arguments{
\item{...}{element a configurer TOKEN et PROXY}
}
\value{
configure les token et proxy pour les acces speciaux
}
\description{
configure
}
\examples{
configure(TOKEN="my_token")
}
