% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_conso_espace.R
\name{conso_enaf.departements}
\alias{conso_enaf.departements}
\title{Retourne les indicateurs annuels de consommation d’espace departementaux}
\usage{
conso_enaf.departements(coddep, ...)
}
\arguments{
\item{coddep}{Code INSEE du département (possibilité de passer un vecteur de code insee sans limite maximum)}

\item{...}{pour futurs developpements}

\item{annee_max}{Année jusqu'à laquelle renvoyer les indicateurs d'artificialisation (incluse)}

\item{annee_min}{Année à partir de laquelle renvoyer les indicateurs d'artificialisation (incluse)}

\item{ordering}{Champs à utiliser pour ordonner le résultat. Defaults to None.}
}
\value{
Retourne les indicateurs annuels de consommation d’espace departementaux
}
\description{
Retourne les indicateurs annuels de consommation d’espace departementaux
}
\examples{
conso_enaf.departements(coddep="59")
conso_enaf.departements(coddep=c("59", "62"), annee_max="2015")
}
