% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv3f_restreint.R
\name{dv3f.geomutations}
\alias{dv3f.geomutations}
\title{Retourne les mutations issues de DV3F pour le périmètre demandée sous forme d’un sf}
\usage{
dv3f.geomutations(code_insee = NULL, lon_lat = NULL, in_bbox = NULL, ...)
}
\arguments{
\item{code_insee}{Codes INSEE communaux ou des arrondissements municipaux.}

\item{lon_lat}{Coordonnée du point au sein de la ou des friches renvoyées c(longitude, latitude)}

\item{in_bbox}{Emprise rectangulaire sous la forme d’un vecteur c(longitude_min, latitude_min, longitude_min, latitude_max)}

\item{...}{pour futurs developpements}
}
\value{
Retourne les mutations issues de DV3F pour le périmètre demandée sous forme d’un sf
}
\description{
Retourne les mutations issues de DV3F pour le périmètre demandée sous forme d’un sf
}
