% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv3f_restreint.R
\name{dv3f.mutation}
\alias{dv3f.mutation}
\title{Renvoie la mutation correspondant à l’identifiant idmutation}
\usage{
dv3f.mutation(idmutation = NULL, ...)
}
\arguments{
\item{idmutation}{identifiant de la mutation}

\item{...}{pour futurs developpements}
}
\value{
Renvoie la mutation correspondant à l’identifiant idmutation
}
\description{
Renvoie la mutation correspondant à l’identifiant idmutation
}
