% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv3f_restreint.R
\name{dv3f_mutations_idmutation}
\alias{dv3f_mutations_idmutation}
\title{Mutation issue de DV3F pour l'identifiant de mutation demandé}
\usage{
dv3f_mutations_idmutation(token, idmutation)
}
\arguments{
\item{token}{token}

\item{idmutation}{A unique value identifying this tup}
}
\value{
Retourne la mutation issue de DV3F pour l'identifiant de mutation demandé
}
\description{
Mutation issue de DV3F pour l'identifiant de mutation demandé
}
\examples{
\dontrun{
dv3f_mutations_idmutation(idmutation="9033720",token=token)
}
}
