% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvf_opendata.R
\name{dvf.mutation}
\alias{dvf.mutation}
\title{Renvoi la mutation correspondant à l’identifiant idmutation}
\usage{
dvf.mutation(idmutation = NULL, ...)
}
\arguments{
\item{idmutation}{identifiant de la mutation}

\item{...}{pour futurs developpements}
}
\value{
Renvoi la mutation correspondant à l’identifiant idmutation
}
\description{
Renvoi la mutation correspondant à l’identifiant idmutation
}
