% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvf_opendata.R
\name{dvf_opendata_mutations}
\alias{dvf_opendata_mutations}
\title{Mutations open}
\usage{
dvf_opendata_mutations(
  anneemut = NULL,
  anneemut_max = NULL,
  anneemut_min = NULL,
  code_insee = NULL,
  codtypbien = NULL,
  codtypproa = NULL,
  codtypprov = NULL,
  contains_geom = NULL,
  fields = NULL,
  filtre = NULL,
  idnatmut = NULL,
  in_bbox = NULL,
  sbati_max = NULL,
  sbati_min = NULL,
  segmtab = NULL,
  sterr_max = NULL,
  sterr_min = NULL,
  valeurfonc_max = NULL,
  valeurfonc_min = NULL,
  vefa = NULL
)
}
\arguments{
\item{anneemut}{Annee de la mutation (>= 2010) - cf anneemut}

\item{anneemut_max}{Annee de la mutation maximale - cf anneemut}

\item{anneemut_min}{Annee de la mutation minimale - cf anneemut}

\item{code_insee}{Code INSEE communal ou d'arrondissement municipal (il est possible d'en demander plusieurs (10 maximum), séparés par des virgules, dans le même département)}

\item{codtypbien}{Code(s) de la typologie de bien à sélectionner (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : codtypbien=11,121 - cf codtypbien}

\item{codtypproa}{Code(s) de la typologie de l'acheteur (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : codtypproa=P,F7 - cf codtypproa}

\item{codtypprov}{Code(s) de la typologie du vendeur (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : codtypprov=P,F7 - cf codtypprov}

\item{contains_geom}{Renvoie les entités dont la géometrie contient celle précisée dans le filtre (en format GeoJSON, WKT, HEXEWKB, WKB). Exemple : /?contains_geom={'type':'Point', 'coordinates':[2.17,46.75]}}

\item{fields}{Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.}

\item{filtre}{Code alphanumerique permettant d'exclure des transactions particulières - cf filtre}

\item{idnatmut}{Code(s) de nature de mutation (il est possible d'en demander plusieurs en séparant par une virgule - exemple : idnatmut=1,2 - cf idnatmut}

\item{in_bbox}{Emprise rectangulaire (via Longitude min,Latitude min,Longitude max,Latitude max). L'emprise demandée ne doit pas excéder un carré de 0.02 deg. x 0.02 deg.}

\item{sbati_max}{Surface maximale bâtie vendue - cf sbati}

\item{sbati_min}{Surface minimale bâtie vendue - cf sbati}

\item{segmtab}{Note de segment du terrain à bâtir - cf segmtab}

\item{sterr_max}{Surface maximale de terrain vendue - cf sterr}

\item{sterr_min}{Surface minimale de terrain vendue - cf sterr}

\item{valeurfonc_max}{Valeur foncière maximale (€) - cf valeurfonc}

\item{valeurfonc_min}{Valeur foncière minimale (€) - cf valeurfonc}

\item{vefa}{Vente en l'état futur d'achèvement - cf vefa}
}
\value{
Retourne les mutations issues de DVF+ pour la commune ou l'emprise rectangulaire demandée (paramètre code_insee ou in_bbox obligatoire)
}
\description{
Mutations open
}
\examples{
dvf_opendata_mutations(code_insee=59001)
}
