% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvf_opendata.R
\name{dvf_opendata_mutations_idmutation}
\alias{dvf_opendata_mutations_idmutation}
\title{Mutation issue de DV3F pour l'identifiant de mutation demandé}
\usage{
dvf_opendata_mutations_idmutation(idmutation)
}
\arguments{
\item{idmutation}{A unique value identifying this tup}
}
\value{
Retourne la mutation pour l'identifiant de mutation demandé
}
\description{
Mutation issue de DV3F pour l'identifiant de mutation demandé
}
\examples{
dvf_opendata_mutations_idmutation(idmutation="9033720")
}
