% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff.geoparcelles}
\alias{ff.geoparcelles}
\title{Retourne les parcelles issues des Fichiers fonciers pour le périmètre demandé sous forme d’un geodataframe integrant les contours géométriques}
\usage{
ff.geoparcelles(code_insee = NULL, lon_lat = NULL, in_bbox = NULL, ...)
}
\arguments{
\item{code_insee}{Code INSEE communal ou d'arrondissement municipal (possibilité de passer un vecteur de code insee sans limite maximum)}

\item{lon_lat}{Coordonnée du point au sein de la ou des friches renvoyées c(longitude, latitude).}

\item{in_bbox}{Emprise rectangulaire (via Longitude min,Latitude min,Longitude max,Latitude max). L'emprise demandée ne doit pas excéder un carré de 0.02 deg. x 0.02 deg.}

\item{...}{pour futurs developpements}
}
\value{
Retourne les parcelles issues des Fichiers fonciers pour le périmètre demandé sous forme d’un geodataframe integrant les contours géométriques
}
\description{
Retourne les parcelles issues des Fichiers fonciers pour le périmètre demandé sous forme d’un geodataframe integrant les contours géométriques
}
