% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff.local}
\alias{ff.local}
\title{Renvoie le local correspondant à l’identifiant idlocal}
\usage{
ff.local(idlocal = NULL, ...)
}
\arguments{
\item{idlocal}{identifiant du local}

\item{...}{pour futurs developpements}
}
\value{
Renvoie le local correspondant à l’identifiant idlocal
}
\description{
Renvoie le local correspondant à l’identifiant idlocal
}
