% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff.parcelle}
\alias{ff.parcelle}
\title{Renvoie la parcelle correspondant à l’identifiant idpar}
\usage{
ff.parcelle(idpar = NULL, ...)
}
\arguments{
\item{idpar}{identifiant de la parcelle}

\item{...}{pour futurs developpements}
}
\value{
Renvoie la parcelle correspondant à l’identifiant idpar
}
\description{
Renvoie la parcelle correspondant à l’identifiant idpar
}
