% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff.proprio}
\alias{ff.proprio}
\title{Renvoie le droit de propriété correspondant à l’identifiant idprodroit}
\usage{
ff.proprio(idprodroit = NULL, ...)
}
\arguments{
\item{idprodroit}{identifiant du proprietaire}

\item{...}{pour futurs developpements}
}
\value{
Renvoie le droit de propriété correspondant à l’identifiant idprodroit
}
\description{
Renvoie le droit de propriété correspondant à l’identifiant idprodroit
}
