% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff.proprios}
\alias{ff.proprios}
\title{Retourne les droits de propriétés des Fichiers fonciers pour le périmètre demandé sous forme d’un dataframe}
\usage{
ff.proprios(code_insee = NULL, lon_lat = NULL, in_bbox = NULL, ...)
}
\arguments{
\item{code_insee}{Code INSEE communal ou d'arrondissement municipal (possibilité de passer un vecteur de code insee sans limite maximum)}

\item{lon_lat}{Coordonnée du point au sein de la ou des friches renvoyées c(longitude, latitude).}

\item{in_bbox}{Emprise rectangulaire (via Longitude min,Latitude min,Longitude max,Latitude max). L'emprise demandée ne doit pas excéder un carré de 0.02 deg. x 0.02 deg.}

\item{...}{pour futurs developpements}

\item{ccodro}{Code(s) du droit réel ou particulier (il est possible de spécifier plusieurs valeurs et de séparer par une virgule) - exemple : ccodro=A,B - cf ccodro}

\item{fields}{Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.}

\item{gtoper}{Indicateur de personne physique ou morale - cf gtoper}

\item{idprocpte}{Identifiant de compte communal - cf idprocpte}

\item{locprop}{Localisation généralisée du propriétaire recevant la Taxe Foncière - cf locprop}

\item{ordering}{Champs à utiliser pour ordonner le résultat. Default NULL}

\item{typedroit}{Type de droit : propriétaire ou gestionnaire - cf typedroit}

\item{catpro3}{Chaîne(s) de caractères contenue dans le code de catégorie de propriétaire (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : catpro3=P,F2,A2c - cf catpro3}
}
\value{
Retourne les droits de propriétés des Fichiers fonciers pour le périmètre demandé sous forme d’un dataframe
}
\description{
Retourne les droits de propriétés des Fichiers fonciers pour le périmètre demandé sous forme d’un dataframe
}
