% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff.tup}
\alias{ff.tup}
\title{Renvoie la tup correspondante à l’identifiant idtup}
\usage{
ff.tup(idtup = NULL, ...)
}
\arguments{
\item{idtup}{identifiant de la TUP}

\item{...}{pour futurs developpements}
}
\value{
Renvoie la tup correspondante à l’identifiant idtup
}
\description{
Renvoie la tup correspondante à l’identifiant idtup
}
