% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff_locaux_idlocal}
\alias{ff_locaux_idlocal}
\title{Locaux pour un identifiant local du local demandé}
\usage{
ff_locaux_idlocal(token, idlocal)
}
\arguments{
\item{token}{token}

\item{idlocal}{A unique value identifying this local.}
}
\value{
Retourne le local des Fichiers fonciers pour l'identifiant fiscal du local demandé
}
\description{
Locaux pour un identifiant local du local demandé
}
\examples{
\dontrun{
ff_locaux_idlocal(idlocal="***",token=token)
}
}
