% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff_locaux_idpk}
\alias{ff_locaux_idpk}
\title{Locaux pour un identifiant fiscal du local demandé}
\usage{
ff_locaux_idpk(token, idpk)
}
\arguments{
\item{token}{token}

\item{idpk}{A unique value identifying this local.}
}
\value{
Retourne le local des Fichiers fonciers pour l'identifiant fiscal du local demandé
}
\description{
Locaux pour un identifiant fiscal du local demandé
}
\examples{
\dontrun{
ff_locaux_idpk(idpk="?",token=token)
}
}
