% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff_parcelles}
\alias{ff_parcelles}
\title{Parcelles}
\usage{
ff_parcelles(
  token,
  code_insee = NULL,
  contains_geom = NULL,
  ctpdl = NULL,
  dcntarti_max = NULL,
  dcntarti_min = NULL,
  dcntnaf_max = NULL,
  dcntnaf_min = NULL,
  dcntpa_max = NULL,
  dcntpa_min = NULL,
  fields = NULL,
  idcomtxt = NULL,
  in_bbox = NULL,
  jannatmin_max = NULL,
  jannatmin_min = NULL,
  nlocal_max = NULL,
  nlocal_min = NULL,
  nlogh_max = NULL,
  nlogh_min = NULL,
  slocal_max = NULL,
  slocal_min = NULL,
  sprincp_max = NULL,
  sprincp_min = NULL,
  stoth_max = NULL,
  stoth_min = NULL,
  catpro3 = NULL
)
}
\arguments{
\item{token}{token}

\item{code_insee}{Code INSEE communal ou d'arrondissement municipal (possibilité de passer un vecteur de code insee sans limite maximum)}

\item{contains_geom}{Renvoie les entités dont la géometrie contient celle précisée dans le filtre (en format GeoJSON, WKT, HEXEWKB, WKB). Exemple : /?contains_geom={'type':'Point', 'coordinates':[2.17,46.75]}}

\item{ctpdl}{Type de pdl (type de copropriété) - cf ctpdl}

\item{dcntarti_max}{Surface artificialisée maximale de la parcelle (m2) - cf dcntarti}

\item{dcntarti_min}{Surface artificialisée minimale de la parcelle (m2) - cf dcntarti}

\item{dcntnaf_max}{Surface NAF maximale de la parcelle (m2) - cf dcntnaf}

\item{dcntnaf_min}{Surface NAF minimale de la parcelle (m2) - cf dcntnaf}

\item{dcntpa_max}{Surface maximale de la parcelle (m2) - cf dcntpa}

\item{dcntpa_min}{Surface minimale de la parcelle (m2) - cf dcntpa}

\item{fields}{Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.}

\item{idcomtxt}{Chaine de caractères contenue dans le libellé de la commune - cf idcomtxt}

\item{in_bbox}{Emprise rectangulaire (via Longitude min,Latitude min,Longitude max,Latitude max). L'emprise demandée ne doit pas excéder un carré de 0.02 deg. x 0.02 deg.}

\item{jannatmin_max}{Année maximale de construction du local le plus ancien - cf jannatmin}

\item{jannatmin_min}{Année minimale de construction du local le plus ancien - cf jannatmin}

\item{nlocal_max}{Nombre de locaux maximal sur la parcelle - cf nlocal}

\item{nlocal_min}{Nombre de locaux minimal sur la parcelle - cf nlocal}

\item{nlogh_max}{Nombre de logements maximal sur la parcelle - cf nlogh}

\item{nlogh_min}{Nombre de logements minimal sur la parcelle - cf nlogh}

\item{slocal_max}{Surface maximale des parties d'évaluation (m2) - cf slocal}

\item{slocal_min}{Surface minimale des parties d'évaluation (m2) - cf slocal}

\item{sprincp_max}{Surface maximale des pièces principales professionnelles (m2) - cf sprincp}

\item{sprincp_min}{Surface minimale des pièces principales professionnelles (m2) - cf sprincp}

\item{stoth_max}{Surface maximale des pièces d'habitation (m2) - cf stoth}

\item{stoth_min}{Surface minimale des pièces d'habitation (m2) - cf stoth}

\item{catpro3}{Chaîne(s) de caractères contenue dans le code de catégorie de propriétaire (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : catpro3=P,F2,A2c - cf catpro3}
}
\value{
Retourne les parcelles issues des Fichiers Fonciers pour la commune ou l'emprise rectangulaire demandée (paramètre in_bbox ou code_insee obligatoire)
}
\description{
Parcelles
}
\examples{
\dontrun{
ff_parcelles(code_insee=59001,token=token)
}
}
