% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff_parcelles_idpar}
\alias{ff_parcelles_idpar}
\title{Locaux pour un identifiant fiscal du local demandé}
\usage{
ff_parcelles_idpar(token, idpar)
}
\arguments{
\item{token}{token}

\item{idpar}{A unique value identifying this parcelle.}
}
\value{
Retourne la parcelle issue des Fichiers fonciers pour l'identifiant de parcelle (référence cadastrale) demandé
}
\description{
Locaux pour un identifiant fiscal du local demandé
}
\examples{
\dontrun{
ff_parcelles_idpar(idpar="02722000BW0072",token=token)
}
}
