% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff_parcelles_idprodroit}
\alias{ff_parcelles_idprodroit}
\title{Droit de propriété des Fichiers fonciers pour l'identifiant idprodroit demandé}
\usage{
ff_parcelles_idprodroit(token, idprodroit)
}
\arguments{
\item{token}{token}

\item{idprodroit}{A unique value identifying this proprio}
}
\value{
Retourne le droit de propriété des Fichiers fonciers pour l'identifiant idprodroit demandé
}
\description{
Droit de propriété des Fichiers fonciers pour l'identifiant idprodroit demandé
}
\examples{
\dontrun{
ff_parcelles_idprodroit(idprodroit="59001+0000101",token=token)
}
}
