% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff_proprios}
\alias{ff_proprios}
\title{Propriétaires}
\usage{
ff_proprios(
  token,
  ccodro = NULL,
  code_insee = NULL,
  fields = NULL,
  gtoper = NULL,
  idprocpte = NULL,
  locprop = NULL,
  ordering = NULL,
  typedroit = NULL,
  catpro3 = NULL
)
}
\arguments{
\item{token}{token}

\item{ccodro}{Code(s) du droit réel ou particulier (il est possible de spécifier plusieurs valeurs et de séparer par une virgule) - exemple : ccodro=A,B - cf ccodro}

\item{code_insee}{Code INSEE communal ou d'arrondissement municipal (il est possible d'en demander plusieurs (10 maximum), séparés par des virgules, dans le même département)}

\item{fields}{Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.}

\item{gtoper}{Indicateur de personne physique ou morale - cf gtoper}

\item{idprocpte}{Identifiant de compte communal - cf idprocpte}

\item{locprop}{Localisation généralisée du propriétaire recevant la Taxe Foncière - cf locprop}

\item{ordering}{Which field to use when ordering the results.}

\item{typedroit}{Type de droit : propriétaire ou gestionnaire - cf typedroit}

\item{catpro3}{Chaîne(s) de caractères contenue dans le code de catégorie de propriétaire (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : catpro3=P,F2,A2c - cf catpro3}
}
\value{
Retourne les droits de propriété issus des Fichiers Fonciers pour la commune demandée (paramètre code_insee obligatoire)
}
\description{
Propriétaires
}
\examples{
\dontrun{
ff_proprios(code_insee="59001",token=token)
}
}
