% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff_tups}
\alias{ff_tups}
\title{TUP}
\usage{
ff_tups(
  token,
  code_insee = NULL,
  contains_geom = NULL,
  typetup = NULL,
  fields = NULL,
  in_bbox = NULL,
  catpro3 = NULL
)
}
\arguments{
\item{token}{token}

\item{code_insee}{Code INSEE communal ou d'arrondissement municipal (possibilité de passer un vecteur de code insee sans limite maximum)}

\item{contains_geom}{Renvoie les entités dont la géometrie contient celle précisée dans le filtre (en format GeoJSON, WKT, HEXEWKB, WKB). Exemple : /?contains_geom={'type':'Point', 'coordinates':[2.17,46.75]}}

\item{typetup}{Type de l’entité : SIMPLE, PDLMP ou UF - cf typetup}

\item{fields}{Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.}

\item{in_bbox}{Emprise rectangulaire (via Longitude min,Latitude min,Longitude max,Latitude max). L'emprise demandée ne doit pas excéder un carré de 0.02 deg. x 0.02 deg.}

\item{catpro3}{Chaîne(s) de caractères contenue dans le code de catégorie de propriétaire (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : catpro3=P,F2,A2c - cf catpro3}
}
\value{
Retourne, en GeoJSON, les tup issues des Fichiers Fonciers pour la commune ou l'emprise demandée (paramètre code_insee ou in_bbox obligatoire)
}
\description{
TUP
}
\examples{
\dontrun{
ff_tups(code_insee=59001,token=token)
}
}
