% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff_tups_idetup}
\alias{ff_tups_idetup}
\title{TUP des Fichiers fonciers pour l'identifiant idtup demandé}
\usage{
ff_tups_idetup(token, idtup)
}
\arguments{
\item{token}{token}

\item{idtup}{A unique value identifying this tup}
}
\value{
Retourne la tup des Fichiers fonciers pour l'identifiant idtup demandé
}
\description{
TUP des Fichiers fonciers pour l'identifiant idtup demandé
}
\examples{
\dontrun{
ff_tups_idetup(idtup="uf027220129581",token=token)
}
}
