% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{graph}
\alias{graph}
\title{Générer des graphiques dynamiques}
\usage{
graph(data, x, y, couleur = NULL, titre = NULL, titre_x = NULL, titre_y = NULL)
}
\arguments{
\item{data}{Les données d'entrées}

\item{x}{une variable en abscisse (catégorielle)}

\item{y}{une variable en ordonnée (numérique ou catégorielle)}

\item{couleur}{une variable pour distinguer par couleur (catégorielle)}

\item{titre}{un titre}

\item{titre_x}{un titre pour l'axe des x}

\item{titre_y}{un titre pour l'axe des y}
}
\value{
Retourne un graphique plotly avec les paramètres demandés
}
\description{
Génère un graphique plotly pour l'ensemble des données du package.
}
\examples{

cartofriches_geofriches(code_insee = 59350) \%>\%
 graph(
x = site_statut, # un champ catégoriel
y = site_type,   # un champ catégoriel
couleur = site_statut,
titre = "Mon titre",
titre_x = "Mon axe des X",
titre_y = "Mon axe des y"
)

cartofriches_geofriches(code_insee = 59350) \%>\%
 graph(
x = site_statut, # un champ catégoriel
y = unite_fonciere_surface, # un champ numérique
titre = "",
titre_x = "",
titre_y = ""
)


}
