% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_conso_espace.R
\name{ind_conso_espace_communes_g}
\alias{ind_conso_espace_communes_g}
\title{Consommation d'espace par commune (Graphique)}
\usage{
ind_conso_espace_communes_g(
  code_insee,
  annee_max = NULL,
  annee_min = NULL,
  ordering = NULL,
  page = NULL,
  page_size = NULL,
  affichage = "type",
  legende = TRUE,
  hectare = TRUE
)
}
\arguments{
\item{code_insee}{Code INSEE communal ou d'arrondissement municipal (possibilité de passer un vecteur de code insee sans limite maximum)}

\item{annee_max}{Année jusqu'à laquelle renvoyer les indicateurs d'artificialisation (incluse)}

\item{annee_min}{Année à partir de laquelle renvoyer les indicateurs d'artificialisation (incluse)}

\item{ordering}{Which field to use when ordering the results.}

\item{page}{A page number within the paginated result set.}

\item{page_size}{Number of results to return per page.}

\item{affichage}{'total' ou par 'type'}

\item{legende}{Affichage de la légende True/False}

\item{hectare}{valeur de la consommation en hectares}
}
\value{
Renvoie un graphique au format plotly des indicateurs de consommation d'espace pour la période comprise entre annee_min et annee_max, bornes incluses, à l'échelle communale
}
\description{
Consommation d'espace par commune (Graphique)
}
\examples{
ind_conso_espace_communes_g(59350)
ind_conso_espace_communes_g(
  59001,
  hectare = FALSE,
  affichage = 'total',
  legende = FALSE
)
}
