% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_conso_espace_dep.R
\name{ind_conso_espace_dep}
\alias{ind_conso_espace_dep}
\title{Consommation d'espace par département}
\usage{
ind_conso_espace_dep(
  coddep,
  annee_max = NULL,
  annee_min = NULL,
  ordering = NULL,
  page = NULL,
  page_size = NULL
)
}
\arguments{
\item{coddep}{Code INSEE du département}

\item{annee_max}{Année jusqu'à laquelle renvoyer les indicateurs d'artificialisation (incluse)}

\item{annee_min}{Année à partir de laquelle renvoyer les indicateurs d'artificialisation (incluse)}

\item{ordering}{Which field to use when ordering the results.}

\item{page}{A page number within the paginated result set.}

\item{page_size}{Number of results to return per page.}
}
\value{
Renvoie les indicateurs de consommation d'espace pour la période comprise entre annee_min et annee_max, bornes incluses, à l'échelle départementale
}
\description{
Consommation d'espace par département
}
\examples{
ind_conso_espace_dep(01)
}
