% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_conso_espace.R
\name{ind_conso_espace_dep_g}
\alias{ind_conso_espace_dep_g}
\title{Consommation d'espace par département (Graphique)}
\usage{
ind_conso_espace_dep_g(
  coddep,
  annee_max = NULL,
  annee_min = NULL,
  ordering = NULL,
  page = NULL,
  page_size = NULL,
  legende = TRUE,
  affichage = "type",
  hectare = TRUE
)
}
\arguments{
\item{coddep}{Code INSEE du département}

\item{annee_max}{Année jusqu'à laquelle renvoyer les indicateurs d'artificialisation (incluse)}

\item{annee_min}{Année à partir de laquelle renvoyer les indicateurs d'artificialisation (incluse)}

\item{ordering}{Which field to use when ordering the results.}

\item{page}{A page number within the paginated result set.}

\item{page_size}{Number of results to return per page.}

\item{legende}{Affichage de la légende True/False}

\item{affichage}{'total' ou par 'type'}

\item{hectare}{valeur de la consommation en hectares}
}
\value{
Renvoie les indicateurs de consommation d'espace pour la période comprise entre annee_min et annee_max, bornes incluses, à l'échelle départementale
}
\description{
Consommation d'espace par département (Graphique)
}
\examples{
ind_conso_espace_dep_g(59)
}
