% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_dv3f.R
\name{ind_dv3f_com_annuel}
\alias{ind_dv3f_com_annuel}
\title{Indicateurs annuels DV3F à l'échelle de la commune}
\usage{
ind_dv3f_com_annuel(
  code_insee,
  annee = NULL,
  ordering = NULL,
  page = NULL,
  page_size = NULL
)
}
\arguments{
\item{code_insee}{Code INSEE communal ou d'arrondissement municipal (possibilité de passer un vecteur de code insee sans limite maximum)}

\item{annee}{Année de mutation}

\item{ordering}{Which field to use when ordering the results.}

\item{page}{A page number within the paginated result set.}

\item{page_size}{Number of results to return per page.}
}
\value{
Renvoie les indicateurs annuels DV3F à l'échelle communale
}
\description{
Indicateurs annuels DV3F à l'échelle de la commune
}
\examples{
ind_dv3f_com_annuel('59350')
}
