% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_dv3f.R
\name{ind_dv3f_epci_annuel}
\alias{ind_dv3f_epci_annuel}
\title{Indicateurs annuels DV3F à l'échelle de l'epci}
\usage{
ind_dv3f_epci_annuel(
  code_epci,
  annee = NULL,
  ordering = NULL,
  page = NULL,
  page_size = NULL
)
}
\arguments{
\item{code_epci}{Code INSEE de l'EPCI (possibilité de passer un vecteur de code insee sans limite maximum)}

\item{annee}{Année de mutation}

\item{ordering}{Which field to use when ordering the results.}

\item{page}{A page number within the paginated result set.}

\item{page_size}{Number of results to return per page.}
}
\value{
Renvoie les indicateurs annuels DV3F à l'échelle de l'epci
}
\description{
Indicateurs annuels DV3F à l'échelle de l'epci
}
\examples{
ind_dv3f_epci_annuel('200093201')
}
