% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_dv3f.R
\name{prix.aav}
\alias{prix.aav}
\title{Retourne les indicateurs annuels ou triennaux de prix à l’aire d’attraction de la ville}
\usage{
prix.aav(code_insee, periode = "annuel", annee = NULL, ordering = NULL, ...)
}
\arguments{
\item{code_insee}{Code AAV de l'aire INSEE}

\item{periode}{Prend la valeur « annuel » ou « triennal ». Defaults to annuel.}

\item{annee}{Année de mutation}

\item{ordering}{Champs à utiliser pour ordonner le résultat. Default NULL}

\item{...}{pour futurs developpements}
}
\value{
dataframe: données sur les marchés immobiliers
}
\description{
Retourne les indicateurs annuels ou triennaux de prix à l’aire d’attraction de la ville
}
\examples{
prix.aav(code_insee="001")
prix.aav(code_insee=c("001", "002"), annee="2020")
prix.aav(code_insee=c("001", "002"), periode="triennal")

}
