% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_dv3f.R
\name{prix.communes}
\alias{prix.communes}
\title{Retourne les indicateurs annuels ou triennaux de prix à la commune}
\usage{
prix.communes(
  code_insee,
  periode = "annuel",
  annee = NULL,
  ordering = NULL,
  ...
)
}
\arguments{
\item{code_insee}{Code INSEE communal ou d'arrondissement municipal (possibilité de passer un vecteur de code insee sans limite maximum)}

\item{periode}{Prend la valeur « annuel » ou « triennal ». Defaults to annuel.}

\item{annee}{Année de mutation}

\item{ordering}{Champs à utiliser pour ordonner le résultat. Default NULL}

\item{...}{pour futurs developpements}
}
\value{
dataframe: données sur les marchés immobiliers
}
\description{
Retourne les indicateurs annuels ou triennaux de prix à la commune
}
\examples{
prix.communes(code_insee="59350")
prix.communes(code_insee=c("59350", "59646"), annee="2020")
prix.communes(code_insee=c("59350", "59646"), periode="triennal")
}
