% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_dv3f.R
\name{prix.epci}
\alias{prix.epci}
\title{Retourne les indicateurs annuels ou triennaux de prix à l'EPCI}
\usage{
prix.epci(code_insee, periode = "annuel", ...)
}
\arguments{
\item{code_insee}{Codes des EPCI}

\item{periode}{Prend la valeur « annuel » ou « triennal ». Defaults to annuel.}

\item{...}{pour futurs developpements}

\item{annee}{Année de mutation}

\item{ordering}{Champs à utiliser pour ordonner le résultat. Default NULL}
}
\value{
dataframe: données sur les marchés immobiliers
}
\description{
Retourne les indicateurs annuels ou triennaux de prix à l'EPCI
}
\examples{
prix.epci(code_insee="200093201")
prix.epci(code_insee="200093201", periode="triennal")

}
