% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_dv3f.R
\name{prix.regions}
\alias{prix.regions}
\title{Retourne les indicateurs annuels ou triennaux de prix à la région}
\usage{
prix.regions(code_insee, periode = "annuel", ...)
}
\arguments{
\item{code_insee}{Codes des régions}

\item{periode}{Prend la valeur « annuel » ou « triennal ». Defaults to annuel.}

\item{...}{pour futurs developpements}

\item{annee}{Année de mutation}

\item{ordering}{Champs à utiliser pour ordonner le résultat. Default NULL}
}
\value{
dataframe: données sur les marchés immobiliers
}
\description{
Retourne les indicateurs annuels ou triennaux de prix à la région
}
\examples{
prix.regions(code_insee="32")
prix.regions(code_insee=c("32", "11"), annee="2020")
prix.regions(code_insee=c("32", "11"), periode="triennal")
}
