% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{tableau}
\alias{tableau}
\title{Générer un tableau des données dynamique}
\usage{
tableau(.data, ...)
}
\arguments{
\item{.data}{Données à afficher}

\item{...}{Autres variables de la fonction DT()}
}
\value{
Retourne un tableau au format DT::datatable
}
\description{
Générer un tableau des données dynamique
}
\examples{
cartofriches_geofriches(code_insee = 59350) \%>\%
dplyr::select(id:unite_fonciere_surface) \%>\%
dplyr::arrange(dplyr::desc(unite_fonciere_surface)) \%>\%
tableau()

}
