% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartofriches.R
\name{cartofriches_friches}
\alias{cartofriches_friches}
\title{Retourne les friches issues de Cartofriches pour la commune, le département ou l'emprise rectangulaire demandée}
\usage{
cartofriches_friches(
  coddep = NULL,
  code_insee = NULL,
  fields = NULL,
  ordering = NULL,
  surface_max = NULL,
  surface_min = NULL,
  urba_zone_type = NULL
)
}
\arguments{
\item{coddep}{Code INSEE départemental}

\item{code_insee}{Code INSEE communal ou d'arrondissement municipal (possibilité de passer un vecteur de code insee sans limite maximum)}

\item{fields}{Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.}

\item{ordering}{Which field to use when ordering the results.}

\item{surface_max}{Surface maximale de l'unité foncière}

\item{surface_min}{Surface minimale de l'unité foncière}

\item{urba_zone_type}{Type de zone d'urbanisme}
}
\value{
Retourne les friches issues de Cartofriches pour la commune, le département ou l'emprise rectangulaire demandée (paramètre code_insee, coddep ou in_bbox obligatoire)
}
\description{
Retourne les friches issues de Cartofriches pour la commune, le département ou l'emprise rectangulaire demandée
}
\examples{
cartofriches_friches(code_insee = 59350)
}
