% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_conso_espace.R
\name{conso_enaf.communes}
\alias{conso_enaf.communes}
\title{Retourne les indicateurs annuels de consommation d’espace communaux}
\usage{
conso_enaf.communes(code_insee, ...)
}
\arguments{
\item{code_insee}{Code INSEE communal ou d'arrondissement municipal (possibilité de passer un vecteur de code insee sans limite maximum)}

\item{...}{pour futurs developpements}

\item{annee_max}{Année jusqu'à laquelle renvoyer les indicateurs d'artificialisation (incluse)}

\item{annee_min}{Année à partir de laquelle renvoyer les indicateurs d'artificialisation (incluse)}

\item{ordering}{Champs à utiliser pour ordonner le résultat. Defaults to None.}
}
\value{
Retourne les indicateurs annuels de consommation d’espace communaux
}
\description{
Retourne les indicateurs annuels de consommation d’espace communaux
}
\examples{

conso_enaf.communes(code_insee="59350")
conso_enaf.communes(code_insee=c("59350", "62041"), annee_min="2015")
}
