% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvf_opendata.R
\name{dvf.geomutations}
\alias{dvf.geomutations}
\title{Retourne les mutations issues de DVF+ opendata pour le périmètre demandée sous forme d’un sf}
\usage{
dvf.geomutations(code_insee = NULL, lon_lat = NULL, in_bbox = NULL, ...)
}
\arguments{
\item{code_insee}{Codes INSEE communaux ou des arrondissements municipaux.}

\item{lon_lat}{Coordonnée du point au sein de la ou des friches renvoyées c(longitude, latitude)}

\item{in_bbox}{Emprise rectangulaire sous la forme d’un vecteur c(longitude_min, latitude_min, longitude_min, latitude_max)}

\item{...}{pour futurs developpements}
}
\value{
Retourne les mutations issues de DVF+ opendata pour le périmètre demandée sous forme d’un sf
}
\description{
Retourne les mutations issues de DVF+ opendata pour le périmètre demandée sous forme d’un sf
}
\examples{
dvf.geomutations(code_insee="59001")
dvf.geomutations(in_bbox=c(3, 50, 3.01, 50.01), fields="all")
dvf.geomutations(code_insee=c("59350", "59646"), valeurfonc_min=1000000)
}
