% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_dv3f.R
\name{prix.departements}
\alias{prix.departements}
\title{Retourne les indicateurs annuels ou triennaux de prix au département}
\usage{
prix.departements(
  coddep,
  periode = "annuel",
  annee = NULL,
  ordering = NULL,
  ...
)
}
\arguments{
\item{coddep}{Codes des départements}

\item{periode}{Prend la valeur « annuel » ou « triennal ». Defaults to annuel.}

\item{annee}{Année de mutation}

\item{ordering}{Champs à utiliser pour ordonner le résultat. Default NULL}

\item{...}{pour futurs developpements}
}
\value{
dataframe: données sur les marchés immobiliers
}
\description{
Retourne les indicateurs annuels ou triennaux de prix au département
}
\examples{
prix.departements(coddep="59")
prix.departements(coddep=c("59", "62"), annee="2020")
prix.departements(coddep=c("59", "62"), periode="triennal")
}
